added loop control

+++ src/CommandLineInterface/processtools.c
@@ -690,11 +690,11 @@ int_fast8_t processinfo_CTRLscreen()
 					
 					printw(" C%d", pinfoarray[pindex]->CTRLval );
 
-                    printw(" %02d:%02d:%02d.%03ld",
+                    printw(" %02d:%02d:%02d.%03d",
                            pinfodisp[pindex].createtime_hr,
                            pinfodisp[pindex].createtime_min,
                            pinfodisp[pindex].createtime_sec,
-                           0.000001*pinfodisp[pindex].createtime_ns);
+                           (int) 0.000001*pinfodisp[pindex].createtime_ns);
 
                     printw("  %6d", pinfolist->PIDarray[pindex]);
                     printw("  %12s", pinfoarray[pindex]->tmuxname);
